autoload admin classes
