+++ admin/field.php
@@ -1,0 +1,7 @@
+<?php
+
+interface P2P_Field {
+	function get_title();
+	function render( $p2p_id, $item );
+}
+
